+++ UIKitDZ/UIKitDZ/Controllers/ProfileViewController.swift
@@ -1,0 +1,7 @@
+// ProfileViewController.swift
+// Copyright © RoadMap. All rights reserved.
+
+import UIKit
+
+/// Экран профиля
+final class ProfileViewController: UIViewController {}
