+++ UIKitDZ/UIKitDZ/NotificationLink.swift
@@ -1,0 +1,14 @@
+// NotificationLink.swift
+// Copyright © RoadMap. All rights reserved.
+
+import UIKit
+
+/// Уведомление
+struct NotificationLink {
+    /// Текст уведомления
+    let text: String
+    /// Аватар
+    let avatar: UIImage
+    /// Картинка к уведомлению
+    let picture: UIImage
+}
