--- UIKitDZ/UIKitDZ/ProfileViewController.swift
@@ -1,7 +1,0 @@
-// ProfileViewController.swift
-// Copyright © RoadMap. All rights reserved.
-
-import UIKit
-
-/// pf
-class ProfileViewController: UIViewController {}
