+++ UIKitDZ/UIKitDZ/SceneDelegate.swift
@@ -12,10 +12,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         willConnectTo session: UISceneSession,
         options connectionOptions: UIScene.ConnectionOptions
     ) {
-        guard let windowScene = (scene as? UIWindowScene) else { return }
-        window = UIWindow(windowScene: windowScene)
-        window?.rootViewController = LinkTabBarController()
-        window?.makeKeyAndVisible()
+        setStartScene(scene)
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {
@@ -45,5 +42,12 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Called as the scene transitions from the foreground to the background.
         // Use this method to save data, release shared resources, and store enough scene-specific state information
         // to restore the scene back to its current state.
+    }
+
+    private func setStartScene(_ scene: UIScene) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        window = UIWindow(windowScene: windowScene)
+        window?.rootViewController = LinkTabBarController()
+        window?.makeKeyAndVisible()
     }
 }
