--- UIKitDZ/UIKitDZ/Source.swift
@@ -1,19 +1,0 @@
-// Source.swift
-// Copyright © RoadMap. All rights reserved.
-
-import UIKit
-
-/// Post
-struct Post {
-    let nickname: String
-    let photo: UIImage
-    let pictures: [UIImage]
-    let timeText: String
-}
-
-/// ds
-struct NotificationLink {
-    let text: String
-    let avatar: UIImage
-    let picture: UIImage
-}
